requirements.txt
Flask
Flask-Session
Werkzeug
Flask-SQLAlchemy
psycopg2-binary
